functions/api/review-list-provider/requirements.txt
google-cloud-firestore==1.7.*
flask==1.1.*
